+++ assets/src/apps/authoring/components/PropertyEditor/schemas/screen.ts
@@ -110,10 +110,10 @@ export const screenUiSchema = {
     'ui:ObjectFieldTemplate': CustomFieldTemplate,
     'ui:title': 'Palette',
     backgroundColor: {
-      'ui:widget': ColorPickerWidget
+      'ui:widget': ColorPickerWidget,
     },
     borderColor: {
-      'ui:widget': ColorPickerWidget
+      'ui:widget': ColorPickerWidget,
     },
     borderStyle: { classNames: 'col-6' },
     borderWidth: { classNames: 'col-6' },
@@ -134,26 +134,27 @@ export const transformScreenModeltoSchema = (activity?: IActivity) => {
     const data = activity?.content?.custom;
     const schemaPalette = {
       ...data.palette,
-      borderWidth: `${data.palette.lineThickness ? data.palette.lineThickness + 'px' : '1px'
-        }`,
+      borderWidth: `${data.palette.lineThickness ? data.palette.lineThickness + 'px' : '1px'}`,
       borderRadius: '10px',
       borderStyle: 'solid',
-      borderColor: `rgba(${data.palette.lineColor || data.palette.lineColor === 0
-        ? chroma(data.palette.lineColor).rgb().join(',')
-        : '255, 255, 255'
-        },${data.palette.lineAlpha})`,
-      backgroundColor: `rgba(${data.palette.fillColor || data.palette.fillColor === 0
-        ? chroma(data.palette.fillColor).rgb().join(',')
-        : '255, 255, 255'
-        },${data.palette.fillAlpha})`
-    }
+      borderColor: `rgba(${
+        data.palette.lineColor || data.palette.lineColor === 0
+          ? chroma(data.palette.lineColor).rgb().join(',')
+          : '255, 255, 255'
+      },${data.palette.lineAlpha})`,
+      backgroundColor: `rgba(${
+        data.palette.fillColor || data.palette.fillColor === 0
+          ? chroma(data.palette.fillColor).rgb().join(',')
+          : '255, 255, 255'
+      },${data.palette.fillAlpha})`,
+    };
     return {
       ...data,
       title: activity?.title,
       Size: { width: data.width, height: data.height },
       checkButton: { showCheckBtn: data.showCheckBtn, checkButtonLabel: data.checkButtonLabel },
       max: { maxAttempt: data.maxAttempt, maxScore: data.maxScore },
-      palette: data.palette.useHtmlProps? data.palette : schemaPalette
+      palette: data.palette.useHtmlProps ? data.palette : schemaPalette,
     };
   }
 };
@@ -172,7 +173,7 @@ export const transformScreenSchematoModel = (schema: any) => {
     palette: { ...schema.palette, useHtmlProps: true },
     trapStateScoreScheme: schema.trapStateScoreScheme,
     negativeScoreAllowed: schema.negativeScoreAllowed,
-    screenButton: schema.screenButton
+    screenButton: schema.screenButton,
   };
 };